gcp/extraction-workflow/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
pandas==1.1.5
google-cloud-bigquery==2.6.1
requests==2.24.0
